Move the jump logic out of the JL adapter into the corresponding feature
