+++ src/adapters/jupyterlab/jl_adapter.ts
@@ -58,6 +58,7 @@ export interface IJupyterLabComponentsManager {
     position: IEditorPosition
   ) => FreeTooltip;
   remove_tooltip: () => void;
+  jumper: CodeJumper;
 }
 
 /**
@@ -311,9 +312,6 @@ export abstract class JupyterLabWidgetAdapter
       }
     }).connect(socket);
 
-    connection.on('goTo', locations =>
-      this.handle_jump(locations, virtual_document.id_path)
-    );
     connection.on('error', e => {
       let error: Error = e.length && e.length >= 1 ? e[0] : new Error();
       // TODO: those codes may be specific to my proxy client, need to investigate
@@ -372,76 +370,6 @@ export abstract class JupyterLabWidgetAdapter
     this.widget.context.model.contentChanged.connect(
       this.update_documents.bind(this)
     );
-  }
-
-  handle_jump(locations: lsProtocol.Location[], id_path: string) {
-    let connection = this.connections.get(id_path);
-
-    // TODO: implement selector for multiple locations
-    //  (like when there are multiple definitions or usages)
-    if (locations.length === 0) {
-      console.log('No jump targets found');
-      return;
-    }
-    console.log('Will jump to the first of suggested locations:', locations);
-
-    let location = locations[0];
-
-    let uri: string = decodeURI(location.uri);
-    let current_uri = connection.getDocumentUri();
-
-    let virtual_position = PositionConverter.lsp_to_cm(
-      location.range.start
-    ) as IVirtualPosition;
-
-    if (uri === current_uri) {
-      let editor_index = this.virtual_editor.get_editor_index(virtual_position);
-      // if in current file, transform from the position within virtual document to the editor position:
-      let editor_position = this.virtual_editor.transform_virtual_to_editor(
-        virtual_position
-      );
-      let editor_position_ce = PositionConverter.cm_to_ce(editor_position);
-      console.log(`Jumping to ${editor_index}th editor of ${uri}`);
-      console.log('Jump target within editor:', editor_position_ce);
-      this.jumper.jump({
-        token: {
-          offset: this.jumper.getOffset(editor_position_ce, editor_index),
-          value: ''
-        },
-        index: editor_index
-      });
-    } else {
-      // otherwise there is no virtual document and we expect the returned position to be source position:
-      let source_position_ce = PositionConverter.cm_to_ce(virtual_position);
-      console.log(`Jumping to external file: ${uri}`);
-      console.log('Jump target (source location):', source_position_ce);
-
-      if (uri.startsWith('file://')) {
-        uri = uri.slice(7);
-      }
-
-      let jump_data = {
-        editor_index: 0,
-        line: source_position_ce.line,
-        column: source_position_ce.column
-      };
-
-      // assume that we got a relative path to a file within the project
-      // TODO use is_relative() or something? It would need to be not only compatible
-      //  with different OSes but also with JupyterHub and other platforms.
-      this.jumper.document_manager.services.contents
-        .get(uri, { content: false })
-        .then(() => {
-          this.jumper.global_jump({ uri, ...jump_data }, false);
-        })
-        .catch(() => {
-          // fallback to an absolute location using a symlink (will only work if manually created)
-          this.jumper.global_jump(
-            { uri: '.lsp_symlink/' + uri, ...jump_data },
-            true
-          );
-        });
-    }
   }
 
   create_adapter(
